+++ Extensions/IISWebAppDeploy/Src/Tasks/IISWebAppDeploy/Main.ps1
@@ -8,6 +8,10 @@ param (
     [string]$webDeployParamFile,
     [string]$overRideParams,
     [string]$websiteName,
+    [string]$removeAdditionalFiles,
+    [string]$excludeFilesFromAppData,
+    [string]$takeAppOffline,
+    [string]$additionalArguments,
     [string]$deployInParallel
     )
 
@@ -15,4 +19,4 @@ $env:CURRENT_TASK_ROOTDIR = Split-Path -Parent $MyInvocation.MyCommand.Path
 
 . $env:CURRENT_TASK_ROOTDIR\DeployIISWebApp.ps1
 
-(Main -machinesList $machinesList -adminUserName $adminUserName -adminPassword $adminPassword -winrmProtocol $winrmProtocol -testCertificate $testCertificate -webDeployPackage "$webDeployPackage" -webDeployParamFile "$webDeployParamFile" -overRideParams "$overRideParams" -websiteName "$websiteName" -deployInParallel $deployInParallel)
+(Main -machinesList $machinesList -adminUserName $adminUserName -adminPassword $adminPassword -winrmProtocol $winrmProtocol -testCertificate $testCertificate -webDeployPackage "$webDeployPackage" -webDeployParamFile "$webDeployParamFile" -overRideParams "$overRideParams" -websiteName "$websiteName" -removeAdditionalFiles $removeAdditionalFiles -excludeFilesFromAppData $excludeFilesFromAppData -takeAppOffline $takeAppOffline -additionalArguments $additionalArguments -deployInParallel $deployInParallel)