+++ Extensions/IISWebAppDeploy/Src/Tasks/IISWebAppDeploy/MsDeployOnTargetMachines.ps1
@@ -59,7 +59,11 @@ function Get-MsDeployCmdArgs
     [Parameter(Mandatory=$true)]
     [string]$webDeployPackage,
     [string]$webDeployParamFile,
-    [string]$overRideParams
+    [string]$overRideParams,
+    [string]$removeAdditionalFiles,
+    [string]$excludeFilesFromAppData,
+    [string]$takeAppOffline,
+    [string]$additionalArguments
     )
 
     if(-not ( Test-Path -Path $webDeployPackage))
@@ -88,6 +92,26 @@ function Get-MsDeployCmdArgs
             $msDeployCmdArgs = [string]::Format('{0} -setParam:{1}', $msDeployCmdArgs, $setParam)
         }
     }
+
+    if($removeAdditionalFiles -eq "false")
+    {
+        $msDeployCmdArgs = [string]::Format('{0} -enableRule:DoNotDeleteRule', $msDeployCmdArgs)
+    }
+
+    if($takeAppOffline -eq "true")
+    {
+        $msDeployCmdArgs = [string]::Format('{0} -enableRule:AppOffline', $msDeployCmdArgs)
+    }
+
+    if($excludeFilesFromAppData -eq "true")
+    {
+        $msDeployCmdArgs = [string]::Format('{0} -skip:Directory="\\App_Data"', $msDeployCmdArgs)
+    }
+
+    if(-not [string]::IsNullOrWhiteSpace($additionalArguments))
+    {
+        $msDeployCmdArgs = [string]::Format('{0} {1}', $msDeployCmdArgs, $additionalArguments)
+    }
     
     $msDeployCmdArgs = [string]::Format(' -verb:sync -source:package="{0}" {1} -dest:auto -retryAttempts:3 -retryInterval:3000', $webDeployPackage, $msDeployCmdArgs)
     return $msDeployCmdArgs
@@ -98,15 +122,42 @@ function Deploy-Website
     param(
         [string]$webDeployPkg,
         [string]$webDeployParamFile,
-        [string]$overRiderParams
+        [string]$overRiderParams,
+        [string]$removeAdditionalFiles,
+        [string]$excludeFilesFromAppData,
+        [string]$takeAppOffline,
+        [string]$additionalArguments
     )
 
     $msDeployExePath = Get-MsDeployLocation -regKeyPath $MsDeployInstallPathRegKey
-    $msDeployCmdArgs = Get-MsDeployCmdArgs -webDeployPackage $webDeployPkg -webDeployParamFile $webDeployParamFile -overRideParams $overRiderParams
+    $msDeployCmdArgs = Get-MsDeployCmdArgs -webDeployPackage $webDeployPkg -webDeployParamFile $webDeployParamFile -overRideParams $overRiderParams -removeAdditionalFiles $removeAdditionalFiles -excludeFilesFromAppData $excludeFilesFromAppData -takeAppOffline $takeAppOffline -additionalArguments $additionalArguments
 
     $msDeployCmd = "`"$msDeployExePath`" $msDeployCmdArgs"
     Write-Verbose "Deploying website. Running command: $msDeployCmd"
     Run-Command -command $msDeployCmd
+}
+
+function Compute-MsDeploy-SetParams
+{
+    param(
+        [string]$websiteName,
+        [string]$overRideParams
+    )
+
+    Write-Verbose "Computing override params for msdeploy."
+
+    if([string]::IsNullOrWhiteSpace($overRideParams))
+    {
+        Write-Verbose "Adding override params to ensure deployment happens on $websiteName"
+        $overRideParams = [string]::Format('name="IIS Web Application Name",value="{0}"', $websiteName)
+    }
+    elseif(!$overRideParams.Contains("IIS Web Application Name")) 
+    {
+        Write-Verbose "Adding override params to ensure deployment happens on $websiteName"
+        $overRideParams = $overRideParams + [string]::Format('{0}name="IIS Web Application Name",value="{1}"',  [System.Environment]::NewLine, $websiteName)
+    }
+
+    return $overRideParams
 }
 
 function Execute-Main
@@ -114,14 +165,25 @@ function Execute-Main
     param (
         [string]$WebDeployPackage,
         [string]$WebDeployParamFile,
-        [string]$OverRideParams
+        [string]$OverRideParams,
+        [string]$WebsiteName,
+        [string]$RemoveAdditionalFiles,
+        [string]$ExcludeFilesFromAppData,
+        [string]$TakeAppOffline,
+        [string]$AdditionalArguments
         )
 
     Write-Verbose "Entering Execute-Main function"
     Write-Verbose "WebDeployPackage = $WebDeployPackage"
     Write-Verbose "WebDeployParamFile = $WebDeployParamFile"
     Write-Verbose "OverRideParams = $OverRideParams"
+    Write-Verbose "WebsiteName = $WebsiteName"
+    Write-Verbose "RemoveAdditionalFiles = $RemoveAdditionalFiles"
+    Write-Verbose "ExcludeFilesFromAppData = $ExcludeFilesFromAppData"
+    Write-Verbose "TakeAppOffline = $TakeAppOffline"
+    Write-Verbose "AdditionalArguments = $AdditionalArguments"
 
-    Deploy-Website -webDeployPkg $WebDeployPackage -webDeployParamFile $WebDeployParamFile -overRiderParams $OverRideParams
+    $overRideParams = Compute-MsDeploy-SetParams -websiteName $websiteName -overRideParams $overRideParams
+    Deploy-Website -webDeployPkg $WebDeployPackage -webDeployParamFile $WebDeployParamFile -overRiderParams $OverRideParams -websiteName $websiteName -excludeFilesFromAppData $excludeFilesFromAppData -removeAdditionalFiles $removeAdditionalFiles -takeAppOffline $takeAppOffline -additionalArguments $AdditionalArguments
     Write-Verbose "Exiting Execute-Main function"
 }